# api/openapi/org.yaml
openapi: "3.1.0"
info:
  title: Mobpilot Org API
  version: "1.0.0"
  description: |
    Organization management, membership, groups, and roles.

servers:
  - url: http://localhost/v1/org
    description: Local development

security:
  - bearerAuth: []

paths:
  /organizations:
    post:
      operationId: createOrg
      summary: Create a new organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrgRequest"
      responses:
        "201":
          description: Organization created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"

  /organizations/{orgID}:
    get:
      operationId: getOrg
      summary: Get an organization by ID
      parameters:
        - name: orgID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "404":
          $ref: "#/components/responses/NotFound"

  /organizations/{orgID}/members:
    get:
      operationId: listMembers
      summary: List all members of an organization
      parameters:
        - name: orgID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Member list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/OrgMember"

  /organizations/{orgID}/members/invite:
    post:
      operationId: inviteMember
      summary: Invite a user to an organization
      parameters:
        - name: orgID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteMemberRequest"
      responses:
        "201":
          description: Invitation created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  token:
                    type: string

  /invitations/accept:
    post:
      operationId: acceptInvitation
      summary: Accept an invitation to join an organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
      responses:
        "200":
          description: Membership created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrgMember"
        "404":
          $ref: "#/components/responses/NotFound"
        "410":
          description: Invitation expired
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"

  /organizations/{orgID}/members/{userID}:
    delete:
      operationId: removeMember
      summary: Remove a member from an organization
      parameters:
        - name: orgID
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: userID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Member removed

  /organizations/{orgID}/groups:
    post:
      operationId: createGroup
      summary: Create a group in an organization
      parameters:
        - name: orgID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroupRequest"
      responses:
        "201":
          description: Group created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
    get:
      operationId: listGroups
      summary: List all groups in an organization
      parameters:
        - name: orgID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Group list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Group"

  /organizations/{orgID}/roles:
    post:
      operationId: createRole
      summary: Create a role in an organization
      parameters:
        - name: orgID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoleRequest"
      responses:
        "201":
          description: Role created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
    get:
      operationId: listRoles
      summary: List all roles in an organization
      parameters:
        - name: orgID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Role list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Role"

  /roles/{roleID}:
    delete:
      operationId: deleteRole
      summary: Delete a role
      parameters:
        - name: roleID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Role deleted

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Organization:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        owner_user_id:
          type: string
          format: uuid
        plan:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        suspended_at:
          type: string
          format: date-time
          nullable: true

    OrgMember:
      type: object
      properties:
        org_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        role:
          type: string
          enum: [owner, admin, member]
        joined_at:
          type: string
          format: date-time

    Group:
      type: object
      properties:
        id:
          type: string
          format: uuid
        org_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        created_at:
          type: string
          format: date-time

    Role:
      type: object
      properties:
        id:
          type: string
          format: uuid
        org_id:
          type: string
          format: uuid
        name:
          type: string
        permissions:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time

    CreateOrgRequest:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
        slug:
          type: string
          pattern: "^[a-z0-9-]+$"

    InviteMemberRequest:
      type: object
      required: [email, role]
      properties:
        email:
          type: string
          format: email
        role:
          type: string
          enum: [admin, member]

    CreateGroupRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string

    CreateRoleRequest:
      type: object
      required: [name, permissions]
      properties:
        name:
          type: string
        permissions:
          type: array
          items:
            type: string

    ProblemDetail:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string

  responses:
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    Conflict:
      description: Resource already exists
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    UnprocessableEntity:
      description: Validation error
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"